+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * switch definitions of `lower` and `lowerBound`, `upper` and `upperBound`
 * universal `Bounded` and lifted `Enum` for Suspension
+* `(Data.Interval.\\)`
 * change names from `IntervalSet -> Covering` and `IntervalLayer -> Layers`
 
 ## 0.0.7.1 -- 2022-02-13